+++ dbt/macros/worldbank.sql
@@ -1,0 +1,3 @@
+{% macro worldbank_fonte_oficial() %}
+  https://www.worldbank.org/en/research/commodity-markets
+{% endmacro %}